move out test db preparation
